+++ app/src/main/java/com/techcourse/exception/HandlerAdapterNotFoundException.java
@@ -1,0 +1,8 @@
+package com.techcourse.exception;
+
+public class HandlerAdapterNotFoundException extends RuntimeException{
+
+    public HandlerAdapterNotFoundException(String e) {
+        super(e);
+    }
+}
