+++ app/src/main/java/com/techcourse/exception/HandlerNotFoundException.java
@@ -1,0 +1,8 @@
+package com.techcourse.exception;
+
+public class HandlerNotFoundException extends RuntimeException{
+
+    public HandlerNotFoundException(String e) {
+        super(e);
+    }
+}
