+++ app/src/main/java/com/techcourse/exception/UncheckedServletException.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.techcourse.exception;
 
 public class UncheckedServletException extends RuntimeException {
